[FEAT] comment CRUD

+++ 6hatsBE/src/main/java/com/ssafy/sixhats/dao/BoardDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BoardDAO extends JpaRepository<BoardVO, Integer>{
+public interface BoardDAO extends JpaRepository<BoardVO, Long>{
     List<BoardVO> findByBoardType(BoardType boardType);
     //BoardVO findNoticeById(BoardVO boardId);
 }
